+++ scratch.js
@@ -1,8 +1,7 @@
 const spacetime = require('./src/index')
 
-let a = spacetime("2019-03-13T18:00:00.000-05:00");
-console.log('timestamp:', a);
-console.log(a.format('iso'));
-
-let s = spacetime.now('etc/gmt-13')
-console.log(s.time())
+console.log(spacetime('March 10, 2019 23:01:00', 'Canada/Pacific').isDST());
+console.log(spacetime('March 11, 2019 00:01:00', 'Canada/Pacific').isDST());
+console.log(spacetime('March 11, 2019 01:01:00', 'Canada/Pacific').isDST());
+console.log(spacetime('March 11, 2019 02:01:00', 'Canada/Pacific').isDST());
+console.log(spacetime('March 11, 2019 03:01:00', 'Canada/Pacific').isDST());