+++ src/input/index.js
@@ -87,8 +87,9 @@ const parseInput = (s, input, givenTz) => {
     return s
   }
   //little cleanup..
-  input = input.replace(/\b(mon|tues|wed|wednes|thu|thurs|fri|sat|satur|sun)(day)?\b/, '')
-  input = input.trim().replace(/ +/g, ' ')
+  input = input.replace(/\b(mon|tues|wed|wednes|thu|thurs|fri|sat|satur|sun)(day)?\b/i, '')
+  input = input.replace(/,/g, '')
+  input = input.replace(/ +/g, ' ').trim()
   //try some known-words, like 'now'
   if (namedDates.hasOwnProperty(input) === true) {
     s = namedDates[input](s)